+++ R/tm_t_crosstable.R
@@ -8,6 +8,12 @@
 #' @param y (`data_extract_spec` or `list` of multiple `data_extract_spec`)
 #'  Object with all available choices with pre-selected option for variable Y - column values
 #'  \code{data_extract_spec} must not allow multiple selection in this case.
+#' @param basic_table_args (`basic_table_args`) object created by [`teal.devel::basic_table_args()`]
+#'  with settings for the module table.
+#'  For more details see the help vignette:
+#'  `vignette("Custom basic_table arguments module", package = "teal.devel")`
+#'  The argument is merged with options variable `teal.basic_table_args` and default module setup.
+#'
 #' @param show_percentage optional, (`logical`) Whether to show percentages
 #'   (relevant only when `x` is a `factor`). Defaults to `TRUE`.
 #' @param show_total optional, (`logical`) Whether to show total column. Defaults to `TRUE`.
@@ -58,7 +64,8 @@
 #'           multiple = FALSE,
 #'           fixed = FALSE
 #'         )
-#'       )
+#'       ),
+#'      basic_table_args = teal.devel::basic_table_args(main_footer = "NEST PROJECT")
 #'     )
 #'   )
 #' )
@@ -72,7 +79,8 @@ tm_t_crosstable <- function(label = "Cross Table",
                             show_percentage = TRUE,
                             show_total = TRUE,
                             pre_output = NULL,
-                            post_output = NULL) {
+                            post_output = NULL,
+                            basic_table_args = teal.devel::basic_table_args()) {
   logger::log_info("Initializing tm_t_crosstable")
   stop_if_not(
     is_character_single(label),
@@ -87,12 +95,15 @@ tm_t_crosstable <- function(label = "Cross Table",
     )
   )
 
+  utils.nest::stop_if_not(inherits(basic_table_args, "basic_table_args"))
+
   ui_args <- as.list(environment())
 
   server_args <- list(
     label = label,
     x = x,
-    y = y
+    y = y,
+    basic_table_args = basic_table_args
   )
 
   module(
@@ -148,7 +159,7 @@ ui_t_crosstable <- function(id, datasets, x, y, show_percentage, show_total, pre
   )
 }
 
-srv_t_crosstable <- function(input, output, session, datasets, label, x, y) {
+srv_t_crosstable <- function(input, output, session, datasets, label, x, y, basic_table_args) {
   init_chunks()
 
   selector_list <- data_extract_multiple_srv(data_extract = list(x = x, y = y), datasets = datasets)
@@ -234,7 +245,7 @@ srv_t_crosstable <- function(input, output, session, datasets, label, x, y) {
 
     chunks_push(substitute(
       expr = {
-        lyt <- rtables::basic_table() %>%
+        lyt <- basic_tables %>%
           split_call %>%
           rtables::add_colcounts() %>%
           tern::summarize_vars(
@@ -246,6 +257,9 @@ srv_t_crosstable <- function(input, output, session, datasets, label, x, y) {
           )
       },
       env = list(
+        basic_tables = parse_basic_table_args(
+          basic_table_args = resolve_basic_table_args(basic_table_args)
+        ),
         split_call = if (show_total) {
           substitute(
             expr = rtables::split_cols_by(