+++ NEWS.md
@@ -11,6 +11,8 @@
 
 ## Enhancements
 * Added support for logging with the `logger` package and added info level logs upon initialization of a module.
+* Added support for custom arguments for `ggplot2::labs` and `ggplot2::theme` in `tm_a_regression`.
+* Added support for custom arguments for `rtables::basic_table` in `tm_t_crosstable`.
 
 ## Miscellaneous
 * Refactored the defunct `teal.devel::data_extract_input` into its replacement `teal.devel::data_extract_ui`.